% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cegr_convert_lon}
\alias{cegr_convert_lon}
\title{Convert (0, 360) longitudes to (-180, 180)}
\usage{
cegr_convert_lon(lon)
}
\arguments{
\item{lon}{\verb{[dbl]} A vector of longitudes in the (0, 360) domain.}
}
\value{
\verb{[dbl]} The equivalent vector of longitudes in the (-180, 180) domain.
}
\description{
Convert (0, 360) longitudes to (-180, 180)
}
\examples{
cegr_convert_lon(c(150, 210))
}
