% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cegr_vars}
\alias{cegr_vars}
\title{Environmental variables}
\description{
The \code{cegr_vars} object contains the environmental variable ids organized by
product name, variable name, and time scale. For example,
\code{cegr_vars$`Biogeochemistry Ocean Model`$o2$historical} is the variable id
for historical O2 from the Biogeochemistry Ocean Model product. It's a
hierarchical named list so you can use RStudio's tab autocomplete feature to
look up products, variables within products, and time scales within
variables. To try it yourself, type \verb{cegr_vars$} and hit tab.
}
\keyword{data}
