% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_var}
\alias{read_var}
\title{Read an environmental variable}
\usage{
read_var(var_id, lon, lat, t)
}
\arguments{
\item{var_id}{\verb{[chr(1)]} The variable id. Use \code{erd_variables} to look up
variable ids by name and time scale, e.g., \code{erd_variables$o2$historical}.}

\item{lon}{\verb{[dbl(n)]} Longitudes of points to extract. Must be the same
length as \code{lat} and \code{t}, and must fall within the \emph{spatial} domain of the
product containing variable \code{var_id}.}

\item{lat}{\verb{[dbl(n)]} Latitudes of points to extract. Must be the same length
as \code{lon} and \code{t}, and must fall within the \emph{spatial} domain of the product
containing variable \code{var_id}.}

\item{t}{\verb{[dbl(n)]} Times of points to extract. Must be the same length as
\code{lon} and \code{lat}, and must fall within the \emph{temporal} domain of the product
containing variable \code{var_id}.}
}
\value{
\verb{[dbl(n)]} The values of the environmental variable extracted from
the points (\code{lon}, \code{lat}) at the points in time closest to \code{t}.
}
\description{
Read an environmental variable
}
\examples{
\dontrun{
set.seed(1234)
x <- cumsum(c(-125, runif(9, -0.2, 0.2)))
y <- cumsum(c(33, runif(9, -0.2, 0.2)))
t <- as.POSIXct("2020-04-01", tz = "UTC") + (0:9) * 3600 * 24 * 7
read_var(erd_variables$o2$historical, x, y, t)
}
}
